proper recovery after kill of container

+++ run.sh
@@ -2,11 +2,14 @@
 set -eo pipefail
 
 if [ -n "${MOVE_BROWSERS_TO_PATH}" ]; then
-  cp -r /ms-playwright/* "${MOVE_BROWSERS_TO_PATH}"
-  export PLAYWRIGHT_BROWSERS_PATH="${MOVE_BROWSERS_TO_PATH}"
-  rm -rf "/ms-playwright"
+  if ! [ "$(ls -A "$MOVE_BROWSERS_TO_PATH")" ]; then
+    cp -r /ms-playwright/* "${MOVE_BROWSERS_TO_PATH}"
+    export PLAYWRIGHT_BROWSERS_PATH="${MOVE_BROWSERS_TO_PATH}"
+    rm -rf "/ms-playwright"
+  fi
 fi
 
+rm -f /tmp/.X0-lock
 Xvfb :0 -screen 0 1280x900x16 -listen tcp -ac &
 
 # Start a web server on port 8282 so we can view the browser
@@ -30,4 +33,4 @@ else
   node --enable-source-maps --unhandled-rejections=strict --trace-uncaught --trace-warnings dist/main-debug.js
 fi
 
-exec redis-cli shutdown
+# redis-cli shutdown
